+++ src/mpmToolbox/Main.java
@@ -13,7 +13,7 @@ import java.io.IOException;
  * @author Axel Berndt
  */
 public class Main {
-    public static final String version = "0.1.3";
+    public static final String version = "0.1.4";
 
     public static void main(String[] args) {
         // read the application settings from file