+++ src/mpmToolbox/gui/audio/AudioDocumentData.java
@@ -1,25 +1,36 @@
 package mpmToolbox.gui.audio;
 
+import com.alee.api.annotations.NotNull;
 import com.alee.extended.tab.DocumentData;
 import com.alee.laf.panel.WebPanel;
-import mpmToolbox.ProjectData;
+import mpmToolbox.gui.ProjectPane;
 
 /**
- * ...
+ * A custom DocumentData object for the audio analysis component.
  * @author Axel Berndt
  */
 public class AudioDocumentData extends DocumentData<WebPanel> {
-    private final ProjectData parentProject;
+    protected final ProjectPane parent;
+    private final WebPanel audioPanel = new WebPanel();
 
     /**
      * constructor
-     * @param id
-     * @param title
-     * @param parentProject
+     * @param parent
      */
-    public AudioDocumentData(String id, String title, ProjectData parentProject) {
-        super(id, title, new WebPanel());       // this.getComponent() returns the WebPanel
-        this.parentProject = parentProject;
+    public AudioDocumentData(@NotNull ProjectPane parent) {
+        super("Audio", "Audio", null);
+
+        this.setComponent(this.audioPanel);
         this.setClosable(false);
+        this.parent = parent;
+//        this.draw();
+    }
+
+    /**
+     * Get the ProjectPane object that this belongs to.
+     * @return
+     */
+    public ProjectPane getParent() {
+        return this.parent;
     }
 }